requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dataclasses-json==0.6.5
dbus-python==1.2.18
distro==1.7.0
Django==3.2.20
djangorestframework
djongo==1.3.6
dnspython==2.6.1
fasteners==0.14.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.2
greenlet==3.0.3
httplib2==0.20.2
huggingface-hub==0.23.0
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
keyring==23.5.0
langchain-community==0.0.36
langchain-core==0.1.50
langsmith==0.1.54
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
MarkupSafe==2.1.5
marshmallow==3.21.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
orjson==3.10.3
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-decouple==3.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.5.10
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.2.4
sympy==1.12
tenacity==8.2.3
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.40.2
typing_extensions==4.11.0
wadllib==1.3.6
yarl==1.9.4
zipp==1.0.0
